+++ plugins/postcss.ts
@@ -15,10 +15,16 @@ export interface Options {
   /** The list of extensions this plugin applies to */
   extensions: string[];
 
-  /** Custom includes path for `postcss-import` */
+  /**
+   * Custom includes path for `postcss-import`
+   * @default `site.options.includes`
+   */
   includes: string | false;
 
-  /** Plugins to use by postcss */
+  /**
+   * Plugins to use by postcss
+   * @default `[postcssNesting(), autoprefixer()]`
+   */
   plugins: unknown[];
 
   /** Set `false` to remove the default plugins */
@@ -28,8 +34,12 @@ export interface Options {
 // Default options
 export const defaults: Options = {
   extensions: [".css"],
-  includes: false,
-  plugins: [],
+  includes: "",
+  plugins: [
+    // @ts-expect-error: postcss-nesting provides wrong types under node16 module resolution: https://github.com/csstools/postcss-plugins/issues/1031
+    postcssNesting(),
+    autoprefixer(),
+  ],
   useDefaultPlugins: true,
 };