+++ tests/assets/vento/vto-filter.tmpl.js
@@ -1,0 +1,6 @@
+export const title = "vento filter example";
+
+export default function (data, { vto }) {
+  const content = `<h1>{{ title |> toUpperCase }}</h1>`;
+  return vto(content, data);
+}