+++ tests/nav.test.ts
@@ -11,3 +11,15 @@ Deno.test("nav plugin", async (t) => {
   await build(site);
   await assertSiteSnapshot(t, site);
 });
+
+Deno.test("nav plugin with pretty urls disabled", async (t) => {
+  const site = getSite({
+    src: "nav",
+    prettyUrls: false,
+  });
+
+  site.use(nav());
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});